fix: move "publish branch" buthub btn to series header with padding
